Move context manager to pytest fixture for usage in export_to_callback tests and make more as a factory
